refactor 예외 처리
